pyproject.toml
[tool.poetry]
name = "stocks-api"
version = "0.1.0"
description = "Coleção de stocks do mercado financeiro"
authors = ["Alessandro <alessandrohenriqueho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
psycopg2-binary = "^2.9.10"
redis = "^5.2.0"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
django-environ = "^0.11.2"
django-redis = "^5.4.0"
djangorestframework = "^3.15.2"
babel = "^2.16.0"
python-dotenv = "^1.0.1"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"
pytest-cov = "^6.0.0"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
